fix: 修复路由协议栈8个问题 — 竞态、重复统计、无效转发等
- transaction tick 超时时设 state=IDLE 防止迟到 response 双重唤醒
- 移除 frag 层重复 tx 统计（由 router 统一计数）
- 广播转发仅在 ttl>1 时执行，避免发送 ttl=0 的无用帧
- bcast_seq_counter 改为 uint8_t 匹配 seq 字段宽度
- pool free 用指针比较替代 ptrdiff_t<0 的未定义行为
- tick batch 溢出时 async callback 留到下轮处理
- 补充 trans_id=slot_index 语义注释和 seen table 覆盖策略注释

+++ sources/route_stack.h
@@ -13,7 +13,7 @@ typedef struct {
     route_transaction_ctx_t *transaction;   // 可选，NULL = 无请求-响应
 
     uint8_t node_id;
-    uint16_t bcast_seq_counter;
+    uint8_t bcast_seq_counter;
 
     void (*on_recv_cb)(void *ctx, uint8_t src, uint8_t trans_id,
                        const uint8_t *data, uint16_t len);
